requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-filter
channels
channels-redis
gunicorn
python-dotenv
